fwlogs: implement IPv6 protocol parsing
This has still one huge hole and that is that it simply sends info about
first extension header instead of protocol itself. The idea is to deploy
it as it is now and get back to it after we collect some data and see
how often such packets are blocked on firewall. It might be that they
are just not that common and thus we might not need to hurry with
implementation. Another option is that they are pretty common and so
simply we would have to implement them as soon as possible.