fwlogs: parse some configuration options and remove -t
This actually uses configuration file for real. We allow configuration
of socket, topic and nflog group.

The additional change here is the removal of -t argument. The topic
change is available only through long option now. The reason for this is
because whole pipeline depends on default topic and topic change breaks
it. It is kind of an internal option and thus it for sure should not
have convenient short option.