fwlogs: handle receive and package handling errors
The receive of packet can fail. That breaks the loop but it won't print
any info. This adds report about that.

The nflog_handle_packet can fail according to some search on the
Internet. This function is not documented in the official documentation
but is used in example that is part of it and no error is handled there
but that minimum the callback can return error so there is probably a
way how this could fail.
This is tied to the  change in callback implementation. Originally we
signaled error from callback if sending failed. This can happen for
various reasons but in general we do not care about that. We print error
and we should keep going because ti might be just temporally thing. That
removes only one place when callback would signal error. Any other place
is considered unexpected error and thus abort is called there. The
result is that we no longer need callback to signal error and thus this
was removed as well.